requirements.txt
fastapi
uvicorn[standard]
jinja2
google-generativeai
sentence-transformers
faiss-cpu
numpy
python-multipart
